RGY Lights
We are driving down the street and see a green traffic light ahead. Because we know precisely the pattern of this traffic light, we know exactly how long we have before it will turn red. We wish to compute whether we will pass the traffic light before it turns red at our current speed.

Write a program that takes the following floating point numbers as input:

our current speed in meter per seconds (s)
the distance to the light in meters (d)
the time until it turns red in seconds (t)
and displays a message indicating whether we will beat the light. You may assume that the input won't be such that we reach the light at exactly the same time it turns red.